Fix BigDecimal type errors in client-management services

+++ client-management/src/main/java/com/dietapp/clientmanagement/service/ClientService.java
@@ -11,6 +11,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.ZoneOffset;
@@ -135,7 +136,7 @@ public class ClientService {
                 .totalClients((int) totalClients)
                 .activeToday((int) activeToday)
                 .sessionsThisWeek((int) sessionsThisWeek)
-                .monthlyRevenue(monthlyRevenue)
+                .monthlyRevenue(BigDecimal.valueOf(monthlyRevenue))
                 .inactiveClientsCount((int) inactiveCount);
     }
